refactor(sandbox): separate driving a model from being an actor
Phase 1 of putting untrusted code on a tensor-parallel model. Nothing about
behaviour changes; this is the seam that has to exist first.

`SandboxModelDeployment` did two jobs. It was the Ray actor — pool, request
lifecycle, statuses, errors — and it was also the *host* of a sandboxed run:
`MediatorProxy` per worker, the interleaved forward, the control events a worker
can park on. Only the first is about being an actor. The second is what any
process holding a model and a socket has to do, and a tensor-parallel shard is
exactly that: a model, a socket, and no actor around it.

`SandboxDriver` (`sandbox/driver.py`) is the second job with the first removed. It
takes a model, a dtype, and somewhere to put a runner's stdout — no Ray, no
request type, no pool. `SandboxModelDeployment` drops from 424 lines to 135 and
now reads as what it is: acquire a runner, hand it to a driver, translate the
result for the client. The one thing the driver cannot know is where a log line
goes, so that arrives as `on_log`.

Also folds three near-identical socket wrappers into one. The sandbox host's
`Connection`, the runner's, and the tensor-parallel `Channel` each reimplemented
framing, timeouts and close over the same `protocol.py`. They are now one
`Channel` with the two direction-specific codecs as overrides — the host sends a
single encoded value and the runner replies with an event name plus values and
kwargs, and that asymmetry is deliberate, so it stays. Net -286 lines.

Verified on 8xA100 that nothing moved: the untrusted suite is 46 passed / 2
failed / 2 skipped, the same two pre-existing `.grad` failures as before the
refactor; the trusted-vs-untrusted conformance suite is still bit-exact; the full
suite 53 passed.

docs/developing/sandboxed-tensor-parallel-proposal.md is the plan this serves.
The short version of why the seam is worth having on its own: a rank cannot act
as a sandbox host until the host logic is separable from the Ray actor, and that
is true whichever topology the sandboxed tensor-parallel design ends up with.
